AdversarialSearch Class added
I added the basics of the adversarial search class. This isn't fully
functional, but should help give a basis of what the class does. it's
based off of pgs. 163-166 in the book, most importantly pg. 166. Some
things aren't clear yet, such as what determines utility, and the
implementation here isn't complete. There is also no pruning. That is
something we can definitely add.