added alpha-beta pruning
I had to change the original code so that when finding the maximums and
minimums we didn't use >= or <=. This changed the board a bit, but it
still results in a tie. To run TTTgame, uncomment it and comment
TTTABgame. The same if you want to run TTTABgame! Alpha-beta seems to be
MUCH faster, as it should!